[1625] Amended extracted sign in command to its own
